requirements.txt
# Core dependencies
aiohttp>=3.9.0
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.4

# Notifications
aiosmtplib>=3.0.1
python-telegram-bot>=20.7
discord-webhook>=1.3.0

# Rate limiting
aiolimiter>=1.1.0

# Type checking
typing-extensions>=4.9.0

# Logging
structlog>=24.1.0

# Development dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.12.0
mypy>=1.7.0
flake8>=6.1.0
